Fixed removing slack teams.
The SlackAgent was being disabled prior to retrieving the agent from the agent manager. The agent couldn't be retrieved from the manager once it was disabled so the team couldn't be remove. The solution was to store the agent before disabling it so that the team can be removed after the agent is disabled.

It seems wise to disable the agent prior to deleting the last team.

I should probably open a bug that status is not cleared prior to a team being removed.